feat: removed unused cookies utils and teacher evaluation page
The commit removes the unused `cookies.dart` utility file, which was previously used for handling webview cookies, and the entire teacher evaluation page (`evaluation.dart`).
Additionally the teacher evaluation page route was removed from the application and the corresponding buttons from the exam result card.
This is because the feature has been removed.